Combine the publish/unpublish code into a single uaclient class
